Enhance JWT permission claims, invoice detail long ID query, and receipt HTML sanitization

+++ Features/Invoices/Read/GetInvoiceByIdQuery.cs
@@ -2,4 +2,4 @@ using MediatR;
 
 namespace SupermarketSystem.Api.Features.Invoices.Read;
 
-public record GetInvoiceByIdQuery(int Id) : IRequest<InvoiceDto?>;
+public record GetInvoiceByIdQuery(long Id) : IRequest<InvoiceDto?>;
